fix: add image, update contact info
