+++ src/app/page.tsx
@@ -1,12 +1,13 @@
 import BookConsultationButton from './components/inputs/BookConsultationButton'
-import ImageBox from './components/layout/ImageBox'
+import Consultation from '../../public/images/konsultation_med_patient.jpg'
 import TreatmentBox from './components/layout/TreatmentBox'
 import H2 from './components/typography/H2'
 import Card from './components/surfaces/Card'
 import { useTranslations } from 'next-intl'
 import SpaceContainer from './components/layout/SpaceContainer'
 import P from './components/typography/P'
-import HeaderWithFooter from './components/HeaderWithFooter'
+import HeaderWithFooter from './components/surfaces/HeaderWithFooter'
+import Image from 'next/image'
 
 const Page: React.FC = () => {
   const t = useTranslations()
@@ -14,7 +15,7 @@ const Page: React.FC = () => {
     <>
       <HeaderWithFooter />
       <main className="mb-56 flex flex-col">
-        <section className="relative flex h-[calc(100vh-80px)] flex-col items-center justify-center bg-[url('/images/Welcome.png')] bg-cover bg-center bg-no-repeat md:bg-contain">
+        <section className="relative flex h-[calc(100vh-80px)] flex-col items-center justify-center bg-[url('/images/Welcome.png')] bg-cover bg-center bg-no-repeat">
           <div className="absolute bottom-12">
             <BookConsultationButton />
           </div>
@@ -28,9 +29,7 @@ const Page: React.FC = () => {
           paragraphs={[t('aboutBox.paragraph1'), t('aboutBox.paragraph2')]}
         />
 
-        <div className="flex w-full justify-center p-0">
-          <ImageBox />
-        </div>
+        <Image src={Consultation} alt={''} />
 
         <TreatmentBox />
